src/hyprctl.lisp: (requst): fix for lispworks
tldr: iolib does send recv with with-pointer-to-vector-data, which in
lispworks8 uses hcl:with-pinned-objects.

usocket's iolib backend will helpfully (inefficiently) copy data to a
static pinnable buffer when sending but when receiving the onus is on
the receiver to provde a static pinnable buffer.  so we cons up a
static vector for every send.